Fix: update UniversityServiceImpl to use ClubPointsRankingDTO instead of Object[]

+++ src/main/java/com/example/uniclub/dto/response/WalletResponse.java
@@ -15,7 +15,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 public class WalletResponse {
 
     private Long walletId;
-    private Integer balancePoints;
+    private Long balancePoints;
+
     private WalletOwnerTypeEnum ownerType;
 
     // Thông tin cho ví CLB (nếu có)
